+++ packages/ui-prompting/src/utilities/renderUtils.tsx
@@ -1,0 +1,25 @@
+import { UIIcon, UiIcons } from '@sap-ux/ui-components';
+import React from 'react';
+
+type IRenderFunction<P> = (props?: P, defaultRender?: (props?: P) => JSX.Element | null) => JSX.Element | null;
+
+export const onLabelRender = function <P>(
+    tooltipText?: string,
+    props?: P,
+    defaultRender?: (props?: P) => React.ReactElement | null
+): React.ReactElement | null {
+    return (
+        <div className="input__label">
+            {defaultRender ? defaultRender(props) : undefined}
+            {tooltipText && <UIIcon iconName={UiIcons.Info} title={tooltipText} />}
+        </div>
+    );
+};
+
+export const getLabelRenderer = function <P>(title?: string): IRenderFunction<P> | undefined {
+    return title
+        ? (props?: P, defaultRender?: (props?: P) => React.ReactElement | null) => {
+              return onLabelRender(title, props, defaultRender);
+          }
+        : undefined;
+};